// apps/frontend/src/api/generated/model/recommendationStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.2.49
 */

export type RecommendationStatus = (typeof RecommendationStatus)[keyof typeof RecommendationStatus]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RecommendationStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  expired: 'expired',
} as const