// apps/frontend/src/api/generated/model/providerConfigOut.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.2.6
 */
import type { ProviderConfigOutBaseUrl } from './providerConfigOutBaseUrl'
import type { ProviderConfigOutModelName } from './providerConfigOutModelName'

export interface ProviderConfigOut {
  id: string
  household_id: string
  provider: string
  priority: number
  enabled: boolean
  base_url: ProviderConfigOutBaseUrl
  model_name: ProviderConfigOutModelName
  ai_data_sharing: string
  created_at: string
  updated_at: string
}